demos/02-rag-search/requirements.txt
promptflow==1.15.0
promptflow-tools==1.4.0
python-dotenv==1.0.0
requests==2.31.0
